fix: cleanup + better text styles for bigger breakpoints
